ai-backend/tools/web_search.py
"""
Web Search Module

This module provides web scraping and search capabilities for gathering
additional context and information about sports events and players.
"""

from typing import Dict, Any, List, Optional
import logging
import aiohttp
from bs4 import BeautifulSoup

logger = logging.getLogger(__name__)


class WebSearchTool:
    """
    Tool for performing web searches and scraping relevant content.
    """
    
    def __init__(self, user_agent: str = "Sport Scribe Bot 1.0"):
        self.user_agent = user_agent
        self.session: Optional[aiohttp.ClientSession] = None
    
    async def __aenter__(self):
        headers = {"User-Agent": self.user_agent}
        self.session = aiohttp.ClientSession(headers=headers)
        return self
    
    async def __aexit__(self, exc_type, exc_val, exc_tb):
        if self.session:
            await self.session.close()
    
    async def search_news(self, query: str, limit: int = 10) -> List[Dict[str, Any]]:
        """
        Search for sports news articles related to the query.
        
        Args:
            query: Search query
            limit: Maximum number of results to return
            
        Returns:
            List of news article data
        """
        # TODO: Implement news search functionality
        logger.info(f"Searching news for query: {query}")
        return []
    
    async def scrape_article(self, url: str) -> Dict[str, Any]:
        """
        Scrape content from a sports article URL.
        
        Args:
            url: URL of the article to scrape
            
        Returns:
            Dictionary containing article content and metadata
        """
        # TODO: Implement article scraping
        logger.info(f"Scraping article: {url}")
        return {}
    
    async def get_team_social_media(self, team_name: str) -> Dict[str, List[str]]:
        """
        Get recent social media posts related to a team.
        
        Args:
            team_name: Name of the team
            
        Returns:
            Dictionary with social media platform as key and posts as values
        """
        # TODO: Implement social media scraping
        logger.info(f"Getting social media for team: {team_name}")
        return {}


class ContentExtractor:
    """
    Tool for extracting and cleaning content from web pages.
    """
    
    @staticmethod
    def extract_article_text(html: str) -> str:
        """
        Extract clean article text from HTML.
        
        Args:
            html: Raw HTML content
            
        Returns:
            Cleaned article text
        """
        # TODO: Implement content extraction using BeautifulSoup
        soup = BeautifulSoup(html, 'html.parser')
        # Remove scripts and style elements
        for script in soup(["script", "style"]):
            script.decompose()
        
        text = soup.get_text()
        lines = (line.strip() for line in text.splitlines())
        chunks = (phrase.strip() for line in lines for phrase in line.split("  "))
        text = ' '.join(chunk for chunk in chunks if chunk)
        
        return text
    
    @staticmethod
    def extract_metadata(html: str) -> Dict[str, Any]:
        """
        Extract metadata from HTML (title, description, etc.).
        
        Args:
            html: Raw HTML content
            
        Returns:
            Dictionary containing metadata
        """
        # TODO: Implement metadata extraction
        soup = BeautifulSoup(html, 'html.parser')
        metadata = {}
        
        # Extract title
        title_tag = soup.find('title')
        if title_tag:
            metadata['title'] = title_tag.get_text().strip()
        
        # Extract meta description
        description_tag = soup.find('meta', attrs={'name': 'description'})
        if description_tag:
            metadata['description'] = description_tag.get('content', '').strip()
        
        return metadata 