ai-backend/requirements.txt
openai-agents>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
supabase>=2.0.0
fastapi>=0.100.0
chainlit>=0.4.0
uvicorn[standard]>=0.23.0
structlog>=23.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
asyncio-mqtt>=0.13.0 